// ecommerce-store/lib/razorpay.d.ts
declare global {
     interface Window {
          Razorpay: new (options: RazorpayOptions) => RazorpayInstance; // Replace `any` with a proper type if available
     }
}

export interface RazorpayOptions {
     key: string;
     amount: number;
     currency: string;
     name: string;
     description: string;
     order_id: string;
     handler: (response: RazorpayResponse) => void;
     prefill: {
          name: string;
          email: string;
          contact: string;
     };
     notes: {
          address: string;
     };
     theme: {
          color: string;
     };
}

export interface RazorpayResponse {
     razorpay_payment_id: string;
     razorpay_order_id: string;
     razorpay_signature: string;
}